Move Dropdown and PaginationControls to common/
Also aria-hidden'd some icons
