+++ src/components/common/Dropdown/index.tsx
@@ -1,16 +1,16 @@
 import DropdownArrow from '@/public/assets/icons/dropdown-arrows.svg';
-import { useEffect, useState } from 'react';
+import { ReactNode, useEffect, useState } from 'react';
 import styles from './style.module.scss';
 
-interface SortOption {
+interface Option {
   value: string;
   label: string;
 }
 
-interface SortDropdownProps {
+interface DropdownProps {
   name: string;
   ariaLabel: string;
-  options: (SortOption | '---')[];
+  options: (Option | '---')[];
   value: string;
   // Justification for disabling rules: This seems to be a false positive.
   // https://stackoverflow.com/q/63767199/
@@ -18,7 +18,7 @@ interface SortDropdownProps {
   onChange: (value: string) => void;
 }
 
-const SortDropdown = ({ name, ariaLabel, options, value, onChange }: SortDropdownProps) => {
+const Dropdown = ({ name, ariaLabel, options, value, onChange }: DropdownProps) => {
   const [open, setOpen] = useState(false);
 
   useEffect(() => {
@@ -30,6 +30,30 @@ const SortDropdown = ({ name, ariaLabel, options, value, onChange }: SortDropdow
     });
   }, [open]);
 
+  const optionButtons: ReactNode[] = [];
+  let dividers = 0;
+  options.forEach(option => {
+    if (option === '---') {
+      optionButtons.push(<hr key={dividers} />);
+      dividers += 1;
+    } else {
+      optionButtons.push(
+        <button
+          type="button"
+          className={styles.option}
+          onClick={event => {
+            onChange(option.value);
+            setOpen(false);
+            event.stopPropagation();
+          }}
+          key={option.value}
+        >
+          {option.label}
+        </button>
+      );
+    }
+  });
+
   return (
     // Justification for disabling rules: For accessibility reasons, the
     // original <select> should be available for keyboard and screen reader
@@ -38,7 +62,6 @@ const SortDropdown = ({ name, ariaLabel, options, value, onChange }: SortDropdow
     // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions
     <div
       className={styles.dropdownWrapper}
-      aria-label={ariaLabel}
       onClick={e => {
         // Using the keyboard to select an option fires the click event on
         // <select>; prevent it from opening the fake dropdown. The <select> has
@@ -50,7 +73,13 @@ const SortDropdown = ({ name, ariaLabel, options, value, onChange }: SortDropdow
         setOpen(true);
       }}
     >
-      <select name={name} id={name} value={value} onChange={e => onChange(e.currentTarget.value)}>
+      <select
+        name={name}
+        id={name}
+        value={value}
+        onChange={e => onChange(e.currentTarget.value)}
+        aria-label={ariaLabel}
+      >
         {options.map(option =>
           option !== '---' ? (
             <option value={option.value} key={option.value}>
@@ -59,32 +88,10 @@ const SortDropdown = ({ name, ariaLabel, options, value, onChange }: SortDropdow
           ) : null
         )}
       </select>
-      <DropdownArrow className={styles.arrow} />
-      <div className={`${styles.contents} ${open ? '' : styles.closed}`}>
-        {options.map((option, i) =>
-          option !== '---' ? (
-            <button
-              type="button"
-              className={styles.option}
-              onClick={event => {
-                onChange(option.value);
-                setOpen(false);
-                event.stopPropagation();
-              }}
-              key={option.value}
-            >
-              {option.label}
-            </button>
-          ) : (
-            // Justification for disabling rule: This is to allow multiple
-            // dividers.
-            // eslint-disable-next-line react/no-array-index-key
-            <hr key={i} />
-          )
-        )}
-      </div>
+      <DropdownArrow className={styles.arrow} aria-hidden />
+      <div className={`${styles.contents} ${open ? '' : styles.closed}`}>{optionButtons}</div>
     </div>
   );
 };
 
-export default SortDropdown;
+export default Dropdown;
