+++ src/components/common/PaginationControls/style.module.scss
