+++ src/components/common/index.ts
@@ -1,3 +1,5 @@
 export { default as Button } from './Button';
+export { default as Dropdown } from './Dropdown';
+export { default as PaginationControls } from './PaginationControls';
 export { default as SEO } from './SEO';
 export { default as VerticalForm } from './VerticalForm';
