+++ src/components/leaderboard/index.ts
@@ -1,4 +1,2 @@
-/* eslint-disable import/prefer-default-export */
 export { default as LeaderboardRow } from './LeaderboardRow';
-export { default as SortDropdown } from './SortDropdown';
 export { default as TopThreeCard } from './TopThreeCard';
